+++ R.Systems.Template.Api.Web/Controllers/CompanyController.cs
@@ -57,7 +57,7 @@ public class CompanyController : ApiControllerBase
     [HttpGet("{companyId}", Name = "GetCompany")]
     public async Task<IActionResult> GetCompany(
         [FromHeader(Name = Headers.Version)] string? version,
-        int companyId,
+        long companyId,
         CancellationToken cancellationToken
     )
     {
@@ -160,7 +160,7 @@ public class CompanyController : ApiControllerBase
     [HttpPut("{companyId}")]
     public async Task<IActionResult> UpdateCompany(
         [FromHeader(Name = Headers.Version)] string? version,
-        int companyId,
+        long companyId,
         UpdateCompanyRequest request
     )
     {
@@ -185,7 +185,7 @@ public class CompanyController : ApiControllerBase
     [HttpDelete("{companyId}")]
     public async Task<IActionResult> DeleteCompany(
         [FromHeader(Name = Headers.Version)] string? version,
-        int companyId
+        long companyId
     )
     {
         DeleteCompanyCommand command = new()
