+++ R.Systems.Template.Api.Web/Controllers/EmployeeController.cs
@@ -55,7 +55,7 @@ public class EmployeeController : ApiControllerBase
     [HttpGet("{employeeId}", Name = "GetEmployee")]
     public async Task<IActionResult> GetEmployee(
         [FromHeader(Name = Headers.Version)] string? version,
-        int employeeId,
+        long employeeId,
         CancellationToken cancellationToken
     )
     {
@@ -162,7 +162,7 @@ public class EmployeeController : ApiControllerBase
     [HttpPut("{employeeId}")]
     public async Task<IActionResult> UpdateEmployee(
         [FromHeader(Name = Headers.Version)] string? version,
-        int employeeId,
+        long employeeId,
         UpdateEmployeeRequest request
     )
     {
@@ -183,7 +183,7 @@ public class EmployeeController : ApiControllerBase
     [HttpDelete("{employeeId}")]
     public async Task<IActionResult> DeleteEmployee(
         [FromHeader(Name = Headers.Version)] string? version,
-        int employeeId
+        long employeeId
     )
     {
         DeleteEmployeeCommand command = new()
