+++ R.Systems.Template.Api.Web/Models/UpdateEmployeeRequest.cs
@@ -4,5 +4,5 @@ public class UpdateEmployeeRequest
 {
     public string? FirstName { get; init; }
     public string? LastName { get; init; }
-    public int CompanyId { get; init; }
+    public long CompanyId { get; init; }
 }
