+++ R.Systems.Template.Core/Common/Domain/Employee.cs
@@ -2,8 +2,8 @@ namespace R.Systems.Template.Core.Common.Domain;
 
 public class Employee
 {
-    public int EmployeeId { get; set; }
+    public long EmployeeId { get; set; }
     public string FirstName { get; set; } = "";
     public string LastName { get; set; } = "";
-    public int? CompanyId { get; init; }
+    public long? CompanyId { get; init; }
 }
