+++ R.Systems.Template.Core/Companies/Commands/DeleteCompany/IDeleteCompanyRepository.cs
@@ -4,5 +4,5 @@ namespace R.Systems.Template.Core.Companies.Commands.DeleteCompany;
 
 public interface IDeleteCompanyRepository : IVersionedRepository
 {
-    Task DeleteAsync(int companyId);
+    Task DeleteAsync(long companyId);
 }
