+++ R.Systems.Template.Core/Companies/Commands/UpdateCompany/CompanyToUpdate.cs
@@ -2,6 +2,6 @@ namespace R.Systems.Template.Core.Companies.Commands.UpdateCompany;
 
 public class CompanyToUpdate
 {
-    public int CompanyId { get; init; }
+    public long CompanyId { get; init; }
     public string Name { get; set; } = "";
 }
