+++ R.Systems.Template.Core/Employees/Commands/DeleteEmployee/IDeleteEmployeeRepository.cs
@@ -4,5 +4,5 @@ namespace R.Systems.Template.Core.Employees.Commands.DeleteEmployee;
 
 public interface IDeleteEmployeeRepository : IVersionedRepository
 {
-    Task DeleteEmployeeAsync(int employeeId);
+    Task DeleteEmployeeAsync(long employeeId);
 }
