+++ R.Systems.Template.Core/Employees/Commands/UpdateEmployee/EmployeeToUpdate.cs
@@ -2,8 +2,8 @@ namespace R.Systems.Template.Core.Employees.Commands.UpdateEmployee;
 
 public class EmployeeToUpdate
 {
-    public int EmployeeId { get; init; }
+    public long EmployeeId { get; init; }
     public string FirstName { get; set; } = "";
     public string LastName { get; init; } = "";
-    public int CompanyId { get; init; }
+    public long CompanyId { get; init; }
 }
