+++ R.Systems.Template.Core/Employees/Queries/GetEmployee/IGetEmployeeRepository.cs
@@ -5,7 +5,7 @@ namespace R.Systems.Template.Core.Employees.Queries.GetEmployee;
 
 public interface IGetEmployeeRepository : IVersionedRepository
 {
-    Task<Employee?> GetEmployeeAsync(int employeeId, CancellationToken cancellationToken);
+    Task<Employee?> GetEmployeeAsync(long employeeId, CancellationToken cancellationToken);
 
-    Task<Employee?> GetEmployeeAsync(int companyId, int employeeId, CancellationToken cancellationToken);
+    Task<Employee?> GetEmployeeAsync(long companyId, long employeeId, CancellationToken cancellationToken);
 }
