+++ R.Systems.Template.Infrastructure.Db/Common/Entities/EmployeeEntity.cs
@@ -2,9 +2,9 @@ namespace R.Systems.Template.Infrastructure.Db.Common.Entities;
 
 internal class EmployeeEntity
 {
-    public int? Id { get; set; }
+    public long? Id { get; set; }
     public string FirstName { get; set; } = "";
     public string LastName { get; set; } = "";
-    public int? CompanyId { get; set; }
+    public long? CompanyId { get; set; }
     public CompanyEntity? Company { get; set; }
 }
